Revert "Merge branch 'lccl' of https://gitee.com/ascend/mxrec into lccl"
This reverts commit 11c2bdd818775bda5647e607000defa0673ee45f.

+++ src/AccCTR/src/embedding_cache/offset_mapper/offset_mapper.h
@@ -75,7 +75,7 @@ public:
         return *this;
     }
 
-    bool Initialize(uint64_t reserve, uint64_t maxSize = 0)
+    bool Initialize(uint32_t reserve, uint32_t maxSize = 0)
     {
         maxCacheSize = maxSize;
         useLength = 0;
@@ -167,7 +167,7 @@ public:
         return ock::ctr::H_OK;
     }
 
-    uint64_t GetUsage()
+    uint32_t GetUsage()
     {
         return useLength - evictSize;
     }
@@ -301,8 +301,8 @@ public:
     }
 
 private:
-    uint64_t maxCacheSize{};            // HBM可容纳embedding条数
-    uint64_t useLength{};               // HBM存储的embedding条数
+    uint32_t maxCacheSize{};            // HBM可容纳embedding条数
+    uint32_t useLength{};               // HBM存储的embedding条数
     LimitedSet *validPos{};             // HBM中可被换出的位置
     LimitedSet *evictPos{};             // 淘汰出的位置
     std::vector<uint64_t> pos2Key;      // HBM中每个位置对应的key
